Improved folder structure: Separated todoService and analyticsService for better organization.
